+++ core/lattice_lightsheet_core/napari.yaml
