+++ plugin/napari_lattice/_reader.py
