+++ pyrightconfig.json
@@ -1,0 +1,9 @@
+{
+  "reportMissingTypeStubs": false,
+  "reportUnknownVariableType": false,
+  "reportUnknownArgumentType": false,
+  "reportUnknownLambdaType": false,
+  "reportUnknownMemberType": false,
+  "reportUnknownParameterType": false,
+  "reportUntypedFunctionDecorator": false
+}
