+++ setup.cfg
@@ -44,20 +44,22 @@ install_requires =
     dask[distributed]
     magic-class>=0.6.13
     magicgui
-    napari[all]
     pyopencl
     read-roi
     gputools
     pyclesperanto_prototype>=0.20.0
-    napari-aicsimageio>=0.7.2
-    napari-spreadsheet
-    napari-workflows>=0.2.8
-    napari-workflow-inspector
     npy2bdv
     redlionfish
     tifffile #>=2022.8.12
     fsspec>=2022.8.2
     
+[options.extras_require]
+napari =
+    napari[all]
+    napari-aicsimageio>=0.7.2
+    napari-spreadsheet
+    napari-workflows>=0.2.8
+    napari-workflow-inspector
 
 [options.extras_require]
 testing =
